[feature/net-70] 영화 검색 쿼리 개선
